+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.9.20] - 2022-06-15
 
 ### Added
+- Configuration parameter for max objects in a workspace to skip narrative reindex
+- Configuration option to skip narrative reindex for certain workspaces
 - Configuration option for poll timeout
 - Configuration option to skip narrative reindexing